requirements.txt
apache-airflow
pycaret
mlflow
flask
flasgger
pandas
numpy
scikit-learn
requests
psutil
